+++ package/naruse-webpack-runner/src/core/template.ts
@@ -1,0 +1,10 @@
+export class NaruseTemplate  {
+    isSupportRecursive = true;
+    buildTemplate () {
+        return 'template';
+    }
+
+    buildPageTemplate () {
+        return 'page';
+    }
+}